Fix bug where ovito is not being saved correctly
- Also fix rounding bug in Silo class

+++ src/main/java/ar/edu/itba/ss/hourglass/models/Silo.java
@@ -262,7 +262,7 @@ public class Silo implements System<Silo.SiloState> {
      * @return Indicates whether the particle's state should be saved.
      */
     private boolean shouldStoreParticles() {
-        return (actualTime / timeStep) % 600 == 0;
+        return (int) (actualTime / timeStep) % 600 == 0;
     }
 
     /**
